share some more code
