+++ controller/pkg/cli/proxy/cmd.go
@@ -4,9 +4,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/flag"
-	"github.com/agentgateway/agentgateway/controller/pkg/cli/profile"
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/proxy/config"
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/proxy/log"
+	"github.com/agentgateway/agentgateway/controller/pkg/cli/proxy/profile"
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/proxy/trace"
 )
